requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis[hiredis]==5.0.1
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
asyncio-throttle==1.0.2
plivo==4.59.1
dataclasses-json==0.6.3

# Crash-proof dependencies
psutil==6.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 